Rename CHITU_F103 variant
Maple platform platformio-build-stm32f1.py uses the 'board' name, not 'board_build.variant'.
