+++ .github/workflows/test-builds.yml
@@ -90,6 +90,7 @@ jobs:
         - NUCLEO_F767ZI
         - REMRAM_V1
         - BTT_SKR_SE_BX
+        - chitu_f103
 
         # Put lengthy tests last
