+++ buildroot/share/PlatformIO/variants/marlin_CHITU_F103/ld/common.inc
