+++ buildroot/share/PlatformIO/variants/marlin_CHITU_F103/ld/stm32f103ze.ld
