+++ buildroot/share/PlatformIO/variants/marlin_CHITU_F103/wirish/boards.cpp
