+++ buildroot/share/PlatformIO/variants/marlin_CHITU_F103/wirish/start_c.c
